backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pymongo==4.9.1
python-dotenv==1.0.1
